online_education/pyproject.toml
[tool.poetry]
name = "onlina-education"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
mypy = "^1.13.0"
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
djangorestframework = "^3.15.2"
pillow = "^11.0.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.10"
drf-yasg = "^1.21.8"
stripe = "^11.4.1"
celery = "^5.4.0"
eventlet = "^0.39.0"
django-celery-beat = "^2.7.0"
redis = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = '.venv'

[tool.isort]
line_length = 119